removes old files, restructures the layout of src folder
