+++ src/assets/index.js
@@ -1,0 +1,4 @@
+import policeBadge from './police-badge.png';
+import policeHelmet from './police-helmet.png';
+
+export { policeBadge, policeHelmet };
