+++ src/components/common/LoadingComponent.js
@@ -1,6 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
+// Currently not being used
 function LoadingComponent(props) {
   const { message } = props;
