+++ src/components/common/Map/index.js
@@ -1,17 +1,17 @@
 import React, { useEffect, useState } from 'react';
 import 'mapbox-gl/dist/mapbox-gl.css';
 import * as mapboxgl from 'mapbox-gl';
-import statesDB from '../../database/states.json';
+import statesDB from '../../../database/states.json';
 import { Button } from 'antd';
 import usZips from 'us-zips';
 import MapButtons from './MapButtons';
-import incidentsDB from '../../database/data2.json';
-import TwitterPopup from './TwitterPopup';
+import incidentsDB from '../../../database/data2.json';
+import TwitterPopup from '../TwitterPopup';
 import ReactDOM from 'react-dom';
 import { useDispatch, useSelector } from 'react-redux';
-import myImg from '../../assets/police-badge.png';
-import myImg2 from '../../assets/police-helmet.png';
-import { fetchIncidents } from '../../state/actions';
+import { policeBadge, policeHelmet } from '../../../assets/';
+
+import { fetchIncidents } from '../../../state/actions';
 
 const Map = () => {
   // using a NYC API to get dummy data for display on the map
@@ -221,14 +221,14 @@ const Map = () => {
     // upload a custom icon image
     map.loadImage(
       // 'https://upload.wikimedia.org/wikipedia/commons/7/7c/201408_cat.png', // --> sample url
-      myImg,
+      policeBadge,
       function(error, image) {
         if (error) throw error;
-        map.addImage('cat', image);
+        map.addImage('badge', image);
       }
     );
 
-    map.loadImage(myImg2, function(error, image) {
+    map.loadImage(policeHelmet, function(error, image) {
       if (error) throw error;
       map.addImage('helmet', image);
     });
