+++ src/index.js
@@ -1,34 +1,21 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { BrowserRouter as Router, Route } from 'react-router-dom';
 import { applyMiddleware, createStore } from 'redux';
 import { Provider } from 'react-redux';
 
 import 'antd/dist/antd.less';
 
 import reducer from './state/reducers/';
-import { Loading } from './components/common/Loading';
-import NavBar from './components/common/NavBar';
+import { App } from './components/common/App';
 import thunk from 'redux-thunk';
 
 const store = createStore(reducer, applyMiddleware(thunk));
 
 ReactDOM.render(
   <Provider store={store}>
-    <Router>
-      <React.StrictMode>
-        <App />
-      </React.StrictMode>
-    </Router>
+    <React.StrictMode>
+      <App />
+    </React.StrictMode>
   </Provider>,
   document.getElementById('root')
 );
-
-function App() {
-  return (
-    <Router>
-      <NavBar />
-      <Route exact path="/" component={Loading} />
-    </Router>
-  );
-}
