--- src/utils/README.md
@@ -1,6 +1,0 @@
-## Utils
-
-- This directory is for any of the business logic that your application may need to use.
-- Simple one-off _helper_ functions that you may need can be defined in this directory.
-- Any reusable logic or handlers or config files can also be used here.
-- **REMINDER do not** check any sensitive information into git
