--- src/utils/oktaConfig.js
@@ -1,9 +1,0 @@
-const config = {
-  issuer: process.env.REACT_APP_OKTA_ISSUER_URI,
-  redirectUri: window.location.origin + '/implicit/callback',
-  clientId: process.env.REACT_APP_CLIENT_ID,
-  pkce: true,
-  scopes: ['openid', 'email', 'profile'],
-};
-
-export { config };
